requirements.txt
python-telegram-bot[job-queue]==20.7
aiohttp==3.9.1
python-dotenv==1.0.0
pytz==2024.1
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
